+++ python-package/lightgbm/basic.py
@@ -248,12 +248,12 @@ class _InnerPredictor(object):
         self.__is_manage_handle = True
         if model_file is not None:
             """Prediction task"""
-            out_num_iterations = ctypes.c_int64(0)
+            out_num_iterations = ctypes.c_int(0)
             _safe_call(_LIB.LGBM_BoosterCreateFromModelfile(
                 c_str(model_file),
                 ctypes.byref(out_num_iterations),
                 ctypes.byref(self.handle)))
-            out_num_class = ctypes.c_int64(0)
+            out_num_class = ctypes.c_int(0)
             _safe_call(_LIB.LGBM_BoosterGetNumClasses(
                 self.handle,
                 ctypes.byref(out_num_class)))
@@ -262,12 +262,12 @@ class _InnerPredictor(object):
         elif booster_handle is not None:
             self.__is_manage_handle = False
             self.handle = booster_handle
-            out_num_class = ctypes.c_int64(0)
+            out_num_class = ctypes.c_int(0)
             _safe_call(_LIB.LGBM_BoosterGetNumClasses(
                 self.handle,
                 ctypes.byref(out_num_class)))
             self.num_class = out_num_class.value
-            out_num_iterations = ctypes.c_int64(0)
+            out_num_iterations = ctypes.c_int(0)
             _safe_call(_LIB.LGBM_BoosterGetCurrentIteration(
                 self.handle,
                 ctypes.byref(out_num_iterations)))
@@ -320,9 +320,9 @@ class _InnerPredictor(object):
                 _safe_call(_LIB.LGBM_BoosterPredictForFile(
                     self.handle,
                     c_str(data),
-                    int_data_has_header,
-                    predict_type,
-                    num_iteration,
+                    ctypes.c_int(int_data_has_header),
+                    ctypes.c_int(predict_type),
+                    ctypes.c_int(num_iteration),
                     c_str(f.name)))
                 lines = f.readlines()
                 nrow = len(lines)
@@ -364,9 +364,9 @@ class _InnerPredictor(object):
         n_preds = ctypes.c_int64(0)
         _safe_call(_LIB.LGBM_BoosterCalcNumPredict(
             self.handle,
-            nrow,
-            predict_type,
-            num_iteration,
+            ctypes.c_int(nrow),
+            ctypes.c_int(predict_type),
+            ctypes.c_int(num_iteration),
             ctypes.byref(n_preds)))
         return n_preds.value
 
@@ -390,12 +390,12 @@ class _InnerPredictor(object):
         _safe_call(_LIB.LGBM_BoosterPredictForMat(
             self.handle,
             ptr_data,
-            type_ptr_data,
-            mat.shape[0],
-            mat.shape[1],
-            C_API_IS_ROW_MAJOR,
-            predict_type,
-            num_iteration,
+            ctypes.c_int(type_ptr_data),
+            ctypes.c_int(mat.shape[0]),
+            ctypes.c_int(mat.shape[1]),
+            ctypes.c_int(C_API_IS_ROW_MAJOR),
+            ctypes.c_int(predict_type),
+            ctypes.c_int(num_iteration),
             ctypes.byref(out_num_preds),
             preds.ctypes.data_as(ctypes.POINTER(ctypes.c_double))))
         if n_preds != out_num_preds.value:
@@ -417,15 +417,15 @@ class _InnerPredictor(object):
         _safe_call(_LIB.LGBM_BoosterPredictForCSR(
             self.handle,
             ptr_indptr,
-            type_ptr_indptr,
+            ctypes.c_int32(type_ptr_indptr),
             csr.indices.ctypes.data_as(ctypes.POINTER(ctypes.c_int32)),
             ptr_data,
-            type_ptr_data,
-            len(csr.indptr),
-            len(csr.data),
-            csr.shape[1],
-            predict_type,
-            num_iteration,
+            ctypes.c_int(type_ptr_data),
+            ctypes.c_int64(len(csr.indptr)),
+            ctypes.c_int64(len(csr.data)),
+            ctypes.c_int64(csr.shape[1]),
+            ctypes.c_int(predict_type),
+            ctypes.c_int(num_iteration),
             ctypes.byref(out_num_preds),
             preds.ctypes.data_as(ctypes.POINTER(ctypes.c_double))))
         if n_preds != out_num_preds.value:
@@ -447,15 +447,15 @@ class _InnerPredictor(object):
         _safe_call(_LIB.LGBM_BoosterPredictForCSC(
             self.handle,
             ptr_indptr,
-            type_ptr_indptr,
+            ctypes.c_int32(type_ptr_indptr),
             csc.indices.ctypes.data_as(ctypes.POINTER(ctypes.c_int32)),
             ptr_data,
-            type_ptr_data,
-            len(csc.indptr),
-            len(csc.data),
-            csc.shape[0],
-            predict_type,
-            num_iteration,
+            ctypes.c_int(type_ptr_data),
+            ctypes.c_int64(len(csc.indptr)),
+            ctypes.c_int64(len(csc.data)),
+            ctypes.c_int64(csc.shape[0]),
+            ctypes.c_int(predict_type),
+            ctypes.c_int(num_iteration),
             ctypes.byref(out_num_preds),
             preds.ctypes.data_as(ctypes.POINTER(ctypes.c_double))))
         if n_preds != out_num_preds.value:
@@ -660,10 +660,10 @@ class Dataset(object):
         ptr_data, type_ptr_data = c_float_array(data)
         _safe_call(_LIB.LGBM_DatasetCreateFromMat(
             ptr_data,
-            type_ptr_data,
-            mat.shape[0],
-            mat.shape[1],
-            C_API_IS_ROW_MAJOR,
+            ctypes.c_int(type_ptr_data),
+            ctypes.c_int(mat.shape[0]),
+            ctypes.c_int(mat.shape[1]),
+            ctypes.c_int(C_API_IS_ROW_MAJOR),
             c_str(params_str),
             ref_dataset,
             ctypes.byref(self.handle)))
@@ -681,13 +681,13 @@ class Dataset(object):
 
         _safe_call(_LIB.LGBM_DatasetCreateFromCSR(
             ptr_indptr,
-            type_ptr_indptr,
+            ctypes.c_int(type_ptr_indptr),
             csr.indices.ctypes.data_as(ctypes.POINTER(ctypes.c_int32)),
             ptr_data,
-            type_ptr_data,
-            len(csr.indptr),
-            len(csr.data),
-            csr.shape[1],
+            ctypes.c_int(type_ptr_data),
+            ctypes.c_int64(len(csr.indptr)),
+            ctypes.c_int64(len(csr.data)),
+            ctypes.c_int64(csr.shape[1]),
             c_str(params_str),
             ref_dataset,
             ctypes.byref(self.handle)))
@@ -705,13 +705,13 @@ class Dataset(object):
 
         _safe_call(_LIB.LGBM_DatasetCreateFromCSC(
             ptr_indptr,
-            type_ptr_indptr,
+            ctypes.c_int(type_ptr_indptr),
             csc.indices.ctypes.data_as(ctypes.POINTER(ctypes.c_int32)),
             ptr_data,
-            type_ptr_data,
-            len(csc.indptr),
-            len(csc.data),
-            csc.shape[0],
+            ctypes.c_int(type_ptr_data),
+            ctypes.c_int64(len(csc.indptr)),
+            ctypes.c_int64(len(csc.data)),
+            ctypes.c_int64(csc.shape[0]),
             c_str(params_str),
             ref_dataset,
             ctypes.byref(self.handle)))
@@ -734,7 +734,7 @@ class Dataset(object):
                     _safe_call(_LIB.LGBM_DatasetGetSubset(
                         handle,
                         used_indices.ctypes.data_as(ctypes.POINTER(ctypes.c_int32)),
-                        used_indices.shape[0],
+                        ctypes.c_int(used_indices.shape[0]),
                         c_str(params_str),
                         ctypes.byref(self.handle)))
                     if self.get_label() is None:
@@ -830,8 +830,8 @@ class Dataset(object):
                 self.handle,
                 c_str(field_name),
                 None,
-                0,
-                FIELD_TYPE_MAPPER[field_name]))
+                ctypes.c_int(0),
+                ctypes.c_int(FIELD_TYPE_MAPPER[field_name])))
             return
         dtype = np.int32 if field_name == 'group' else np.float32
         data = list_to_1d_numpy(data, dtype, name=field_name)
@@ -849,8 +849,8 @@ class Dataset(object):
             self.handle,
             c_str(field_name),
             ptr_data,
-            len(data),
-            type_data))
+            ctypes.c_int(len(data)),
+            ctypes.c_int(type_data)))
 
     def get_field(self, field_name):
         """Get property from the Dataset.
@@ -865,8 +865,8 @@ class Dataset(object):
         info : array
             A numpy array of information of the data
         """
-        tmp_out_len = ctypes.c_int64()
-        out_type = ctypes.c_int32()
+        tmp_out_len = ctypes.c_int()
+        out_type = ctypes.c_int()
         ret = ctypes.POINTER(ctypes.c_void_p)()
         _safe_call(_LIB.LGBM_DatasetGetField(
             self.handle,
@@ -955,7 +955,7 @@ class Dataset(object):
             _safe_call(_LIB.LGBM_DatasetSetFeatureNames(
                 self.handle,
                 c_array(ctypes.c_char_p, c_feature_name),
-                len(feature_name)))
+                ctypes.c_int(len(feature_name))))
 
     def set_label(self, label):
         """
@@ -1076,7 +1076,7 @@ class Dataset(object):
         number of rows : int
         """
         if self._is_constructed:
-            ret = ctypes.c_int64()
+            ret = ctypes.c_int()
             _safe_call(_LIB.LGBM_DatasetGetNumData(self.handle,
                                                    ctypes.byref(ret)))
             return ret.value
@@ -1092,7 +1092,7 @@ class Dataset(object):
         number of columns : int
         """
         if self._is_constructed:
-            ret = ctypes.c_int64()
+            ret = ctypes.c_int()
             _safe_call(_LIB.LGBM_DatasetGetNumFeature(self.handle,
                                                       ctypes.byref(ret)))
             return ret.value
@@ -1147,7 +1147,7 @@ class Booster(object):
                 _safe_call(_LIB.LGBM_BoosterMerge(
                     self.handle,
                     self.__init_predictor.handle))
-            out_num_class = ctypes.c_int64(0)
+            out_num_class = ctypes.c_int(0)
             _safe_call(_LIB.LGBM_BoosterGetNumClasses(
                 self.handle,
                 ctypes.byref(out_num_class)))
@@ -1158,12 +1158,12 @@ class Booster(object):
             self.__get_eval_info()
         elif model_file is not None:
             """Prediction task"""
-            out_num_iterations = ctypes.c_int64(0)
+            out_num_iterations = ctypes.c_int(0)
             _safe_call(_LIB.LGBM_BoosterCreateFromModelfile(
                 c_str(model_file),
                 ctypes.byref(out_num_iterations),
                 ctypes.byref(self.handle)))
-            out_num_class = ctypes.c_int64(0)
+            out_num_class = ctypes.c_int(0)
             _safe_call(_LIB.LGBM_BoosterGetNumClasses(
                 self.handle,
                 ctypes.byref(out_num_class)))
@@ -1198,7 +1198,7 @@ class Booster(object):
         model = state['handle']
         if model is not None:
             handle = ctypes.c_void_p()
-            out_num_iterations = ctypes.c_int64(0)
+            out_num_iterations = ctypes.c_int(0)
             with _temp_file() as f:
                 f.writelines(model)
                 _safe_call(_LIB.LGBM_BoosterCreateFromModelfile(
@@ -1335,7 +1335,7 @@ class Booster(object):
         self.__is_predicted_cur_iter = [False for _ in range(self.__num_dataset)]
 
     def current_iteration(self):
-        out_cur_iter = ctypes.c_int64(0)
+        out_cur_iter = ctypes.c_int(0)
         _safe_call(_LIB.LGBM_BoosterGetCurrentIteration(
             self.handle,
             ctypes.byref(out_cur_iter)))
@@ -1422,7 +1422,7 @@ class Booster(object):
             num_iteration = self.best_iteration
         _safe_call(_LIB.LGBM_BoosterSaveModel(
             self.handle,
-            num_iteration,
+            ctypes.c_int(num_iteration),
             c_str(filename)))
 
     def dump_model(self, num_iteration=-1):
@@ -1441,13 +1441,13 @@ class Booster(object):
         if num_iteration <= 0:
             num_iteration = self.best_iteration
         buffer_len = 1 << 20
-        tmp_out_len = ctypes.c_int64(0)
+        tmp_out_len = ctypes.c_int(0)
         string_buffer = ctypes.create_string_buffer(buffer_len)
         ptr_string_buffer = ctypes.c_char_p(*[ctypes.addressof(string_buffer)])
         _safe_call(_LIB.LGBM_BoosterDumpModel(
             self.handle,
-            num_iteration,
-            buffer_len,
+            ctypes.c_int(num_iteration),
+            ctypes.c_int(buffer_len),
             ctypes.byref(tmp_out_len),
             ptr_string_buffer))
         actual_len = tmp_out_len.value
@@ -1457,8 +1457,8 @@ class Booster(object):
             ptr_string_buffer = ctypes.c_char_p(*[ctypes.addressof(string_buffer)])
             _safe_call(_LIB.LGBM_BoosterDumpModel(
                 self.handle,
-                num_iteration,
-                actual_len,
+                ctypes.c_int(num_iteration),
+                ctypes.c_int(actual_len),
                 ctypes.byref(tmp_out_len),
                 ptr_string_buffer))
         return json.loads(string_buffer.value.decode())
@@ -1533,10 +1533,10 @@ class Booster(object):
         ret = []
         if self.__num_inner_eval > 0:
             result = np.array([0.0 for _ in range(self.__num_inner_eval)], dtype=np.float64)
-            tmp_out_len = ctypes.c_int64(0)
+            tmp_out_len = ctypes.c_int(0)
             _safe_call(_LIB.LGBM_BoosterGetEval(
                 self.handle,
-                data_idx,
+                ctypes.c_int(data_idx),
                 ctypes.byref(tmp_out_len),
                 result.ctypes.data_as(ctypes.POINTER(ctypes.c_double))))
             if tmp_out_len.value != self.__num_inner_eval:
@@ -1576,7 +1576,7 @@ class Booster(object):
             data_ptr = self.__inner_predict_buffer[data_idx].ctypes.data_as(ctypes.POINTER(ctypes.c_double))
             _safe_call(_LIB.LGBM_BoosterGetPredict(
                 self.handle,
-                data_idx,
+                ctypes.c_int(data_idx),
                 ctypes.byref(tmp_out_len),
                 data_ptr))
             if tmp_out_len.value != len(self.__inner_predict_buffer[data_idx]):
@@ -1590,7 +1590,7 @@ class Booster(object):
         """
         if self.__need_reload_eval_info:
             self.__need_reload_eval_info = False
-            out_num_eval = ctypes.c_int64(0)
+            out_num_eval = ctypes.c_int(0)
             """Get num of inner evals"""
             _safe_call(_LIB.LGBM_BoosterGetEvalCounts(
                 self.handle,
@@ -1598,7 +1598,7 @@ class Booster(object):
             self.__num_inner_eval = out_num_eval.value
             if self.__num_inner_eval > 0:
                 """Get name of evals"""
-                tmp_out_len = ctypes.c_int64(0)
+                tmp_out_len = ctypes.c_int(0)
                 string_buffers = [ctypes.create_string_buffer(255) for i in range(self.__num_inner_eval)]
                 ptr_string_buffers = (ctypes.c_char_p * self.__num_inner_eval)(*map(ctypes.addressof, string_buffers))
                 _safe_call(_LIB.LGBM_BoosterGetEvalNames(
